feat: extract settings page to separate component
